Initial flat directory refactor
